+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Unknown browsers now return `:unknown_browser` as the id
 - All the changes above affect how `browser.meta` is composed
 - Add method `Browser::Base#unknown?`
+- Fix issue with `Browser::Base#safari?` matching full version
 
 ## 4.2.0